+++ hata/discord/events/event_types.py
@@ -1,4 +1,4 @@
-__all__ = ('GuildUserChunkEvent', )
+__all__ = ('GuildUserChunkEvent', 'VoiceServerUpdateEvent',)
 
 from ..bases import EventBase
 from ..user import User
@@ -43,3 +43,52 @@ class GuildUserChunkEvent(EventBase):
         yield self.nonce
         yield self.index
         yield self.count
+
+
+class VoiceServerUpdateEvent(EventBase):
+    """
+    Represents an `VOICE_SERVER_UPDATE` event.
+    
+    Attributes
+    ----------
+    endpoint : `None` or `str`
+        The voice server's host.
+    guild_id : `int`
+        The respective guild's identifier.
+    token : `str`
+        Voice connection token.
+    """
+    __slots__ = ('endpoint', 'guild_id', 'token')
+    
+    def __repr__(self):
+        """Returns the representation of the voice server update event."""
+        repr_parts = ['<', self.__class__.__name__,]
+        
+        endpoint = self.endpoint
+        if (endpoint is not None):
+            repr_parts.append(' endpoint=')
+            repr_parts.append(repr(endpoint))
+            repr_parts.append(', ')
+        
+        repr_parts.append(' guild_id=')
+        repr_parts.append(repr(self.guild_id))
+        
+        repr_parts.append(', token=')
+        repr_parts.append(repr(self.token))
+        
+        repr_parts.append('>')
+        return ''.join(repr_parts)
+        
+    def __len__(self):
+        """Helper for unpacking if needed."""
+        return 3
+    
+    def __iter__(self):
+        """
+        Unpacks the voice server update event.
+        
+        This method is a generator.
+        """
+        yield self.endpoint
+        yield self.guild_id
+        yield self.token